Fixed bugs in redirect after adding a new person
When adding a new person the user would be redirected to the main person search
page which would display the user they just created.  They would not be able
to modify the search field though as it would generate errors from the AJAX
call because the AJAX was coded with a relative search path.  This commit
fixes that problem by absolutely referencing that search URL.  It also
further clarifity to the search results by populating the search field with
a search for the user they just created.