Fixes a search bug that causes a hard crash
When multiple users have the same first/last name the _search form adds
their email to the display inside of parentheses.  The email was being pulled
from person.user.email which doesn't exist for a person that is not a user since
person.user is nil.  The email should come from person.email.