Enable pages to use JavaScript to populate notice and error messages
The notice and error divs right now only appear during page load if there
was a message to display.  There are times when a page might want to pop
 up a message to the user and use the formatting of the notice and error divs.

This commit makes sure the error and notice divs are always present on every page.
The use just needs to set the message and show the div.