Moved ajax calls to household screen from person page to their own view/controller
The ajax calls to select or create a household on the person page were directly
calling the household#index page which container buttons that assumed clicking on
a household would update a person object on the main screens document.

We have decided we need a fully manageable household page, so we need to take this
person specific logic for selecting a household and put it in its own view/controller.